git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@5902 35acf78f-673a-0410-8e92-d51de3d6d3f4

+++ os/kernel/include/chvt.h
@@ -44,7 +44,7 @@
  * @api
  */
 #define S2ST(sec)                                                           \
-  ((systime_t)(((uint32_t)(sec)) * ((uint32_t)CH_FREQUENCY)))
+  ((systime_t)((sec) * CH_FREQUENCY))
 
 /**
  * @brief   Milliseconds to system ticks.
